requirements.txt
numpy==1.25.2
pandas==2.2.3
scipy==1.15.1
python-dateutil==2.9.0.post0
pytz==2024.2
tzdata==2025.1
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchvision==0.16.1+cu121
triton==2.1.0
SimpleITK==2.4.0
nibabel==5.3.2
pydicom==3.0.1
scikit-image==0.25.0
opencv-python==4.8.0.76
scikit-learn==1.6.1
joblib==1.4.2
threadpoolctl==3.5.0
pytorch-tabnet==4.1.0
matplotlib==3.10.0
seaborn>=0.13.0  
tensorboard==2.18.0
tensorboardX==2.6.2.2
tqdm==4.67.1
rich==14.1.0
Pillow==10.2.0
PyYAML==6.0.2
requests==2.28.1
typing_extensions==4.14.0
packaging==24.2
psutil==7.0.0